fix: git conflicts

+++ src/main/java/com/anthony/calorie_counter/service/impl/UserService.java
@@ -4,36 +4,43 @@ import com.anthony.calorie_counter.entity.User;
 import com.anthony.calorie_counter.exceptions.NotFoundException;
 import com.anthony.calorie_counter.repository.UserRepository;
 import com.anthony.calorie_counter.service.IUserService;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 @Service
 public class UserService implements IUserService {
     @Autowired
     UserRepository userRepository;
 
-    @Override
+    @Override @Transactional(readOnly = true)
     public User findById(Long id) {
         return userRepository.findById(id)
                 .orElseThrow(() -> new NotFoundException("User %d was not found.".formatted(id)));
     }
 
-    @Override
+    @Override @Transactional
     public User save(User user) {
         return userRepository.save(user);
     }
 
-    @Override
-    public User update(Long id, User user) {
-        findById(id);
-        user.setId(id);
+    @Override @Transactional
+    public User update(Long id, User updateUser) {
+        try {
+        User user = userRepository.getReferenceById(id);
+        user.setFullName(updateUser.getFullName());
+        user.setEmail(updateUser.getEmail());
+        user.setPassword(updateUser.getPassword());
         return userRepository.save(user);
+        } catch (EntityNotFoundException e) {
+            throw new NotFoundException("User %d was not found.".formatted(id));
+        }
     }
 
     @Override
     public void delete(Long id) {
-        findById(id);
+        if (!userRepository.existsById(id)) throw new NotFoundException("User %d was not found.".formatted(id));
         userRepository.deleteById(id);
     }
-
 }
